FEBio2: Fixed bug in FETet10_::shape_deriv2. Addressed several compiler warnings. Made some enhancements to FEBioPlotFile. Removed FEModel parameter to FEDomain::Initialize. The FEMesh now initializes the domains (instead of FEModel).
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@8498 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBio2/CommandManager.h
@@ -16,11 +16,11 @@ public:
 
 public:
 	void AddCommand(Command* pcmd) { m_Cmd.push_back(pcmd); }
-	int Size() { return m_Cmd.size(); }
+	int Size() { return (int)m_Cmd.size(); }
 
 	Command* Find(const char* szcmd)
 	{
-		int N = m_Cmd.size();
+		int N = (int)m_Cmd.size();
 		if (N == 0) return 0;
 
 		list<Command*>::iterator ic = m_Cmd.begin();
